classe de exceções criadas

+++ src/main/java/com/projeto_estudo_spring/service/ProdutoService.java
@@ -5,6 +5,8 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.projeto_estudo_spring.exception.RecursoNotFoundException;
 import com.projeto_estudo_spring.model.Produto;
 import com.projeto_estudo_spring.repository.ProdutoRepository;
 
@@ -22,12 +24,19 @@ public class ProdutoService {
         return repository.findAll();
     }
 
-    public Optional<Produto>buscarPorId(Long id){
-        return repository.findById(id);
+    public Produto buscarPorId(Long id){
+        return repository.findById(id)
+        .orElseThrow(() -> new RecursoNotFoundException("Produto com Id " +id+" não encontrado"));
     }
 
     public void deletarProduto(Long id){
-        repository.deleteById(id);
+
+        if (!repository.existsById(id)) {
+            throw new RecursoNotFoundException("Produto com Id " +id+ " não encontrado");
+        }else{
+            repository.deleteById(id);
+        }
+        
     }
